Inventory_management_system/pyproject.toml
[project]
name = "inventory-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "load-dotenv>=0.1.0",
    "pwinput>=1.0.3",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
]
